ok, let’s try building ffmpeg on the fly
